+++ engine/utils/src/test/java/org/asqatasun/utils/colornamelookup/ColorNameLookupTest.java
@@ -22,7 +22,7 @@ package org.asqatasun.utils.colornamelookup;
 
 import java.awt.Color;
 import junit.framework.TestCase;
-import org.junit.Test; // Junit anotation @Test
+import org.junit.Test;  // Junit 4 anotation @Test
 
 /**
  *
@@ -45,6 +45,47 @@ public class ColorNameLookupTest extends TestCase {
 
 
     // @@@TODO add test for getColorNameFromStr
+
+    ///// getColorFromName ///////////////////////////////////////////////
+
+    @Test
+    public void testGetColorNameFromStrWhite() {
+        String colorStr = "white";
+        System.out.println("GetColorNameFromStr ["+ colorStr +"]");
+        ColorNameLookup c = new ColorNameLookup();
+        String result     = c.getColorNameFromStr(colorStr);
+        assertEquals("White", result);
+    }
+
+    @Test
+    public void testGetColorNameFromStrBadColorName() {
+        String colorStr = "BadColorName";
+        System.out.println("GetColorNameFromStr ["+ colorStr +"]");
+        ColorNameLookup c = new ColorNameLookup();
+        String result     = c.getColorNameFromStr(colorStr);
+        assertEquals("", result);
+    }
+
+    @Test
+    public void testGetColorNameFromStrBadCaseWidthSpaces() {
+        String colorStr = " darkmagenta   ";
+        System.out.println("GetColorNameFromStr ["+ colorStr +"]");
+        ColorNameLookup c = new ColorNameLookup();
+        String result     = c.getColorNameFromStr(colorStr);
+        assertEquals("DarkMagenta", result);
+    }
+
+
+    ///// getColorFromName ///////////////////////////////////////////////
+
+    @Test
+    public void testGetColorFromNameBadColorName() {
+        String colorStr = "BadColorName";
+        System.out.println("GetColorFromNameBadColorName ["+ colorStr +"]");
+        ColorNameLookup c = new ColorNameLookup();
+        Color result    = c.getColorFromName(colorStr);
+        assertNull(result);
+    }
 
     @Test
     public void testGetColorFromNameWhite() {
@@ -85,6 +126,5 @@ public class ColorNameLookupTest extends TestCase {
         Color result    = c.getColorFromName(colorStr);
         assertEquals(expResult, result);
     }
-
 
 }